requirements.txt
Django
djangorestframework
django-channels
psycopg2-binary
redis
pytest
pytest-django
websockets
django-redis
channels
channels-redis
pytest-asyncio
celery
dotenv
python-dotenv